+++ packages/cubejs-firebolt-driver/test/autostart.test.ts
@@ -1,5 +1,8 @@
+import { assertDataSource, getEnv } from '@cubejs-backend/shared';
 import { DriverTests } from '@cubejs-backend/testing-shared';
 
+import { Firebolt } from 'firebolt-sdk';
+import { version } from 'firebolt-sdk/package.json';
 import { FireboltDriver } from '../src';
 
 describe('FireboltDriver autostart', () => {
@@ -28,6 +31,40 @@ describe('FireboltDriver autostart', () => {
       await tests.testQuery();
     } catch (error) {
       expect(driver.ensureEngineRunning).toHaveBeenCalled();
+    }
+  });
+  test('starts the engine after connection', async () => {
+    const dataSource = assertDataSource('default');
+
+    const username = getEnv('dbUser', { dataSource });
+    const auth = username.includes('@')
+      ? { username, password: getEnv('dbPass', { dataSource }) }
+      : { client_id: username, client_secret: getEnv('dbPass', { dataSource }) };
+    const engineName = getEnv('fireboltEngineName', { dataSource });
+    const firebolt = Firebolt({
+      apiEndpoint: getEnv('fireboltApiEndpoint', { dataSource }) || 'api.app.firebolt.io',
+    });
+    await firebolt.connect({
+      auth,
+      database: getEnv('dbName', { dataSource }),
+      account: getEnv('fireboltAccount', { dataSource }),
+      engineEndpoint: getEnv('fireboltEngineEndpoint', { dataSource }),
+      additionalParameters: {
+        userClients: [{
+          name: 'CubeDev+Cube',
+          version
+        }]
+      },
+    });
+
+    const engine = await firebolt.resourceManager.engine.getByName(engineName);
+    try {
+      await engine.stop();
+
+      driver = new FireboltDriver({});
+      await driver.testConnection();
+    } finally {
+      await engine.start();
     }
   });
 });