+++ rust/cubesqlplanner/cubesqlplanner/src/planner/base_member.rs
@@ -1,18 +1,24 @@
 use super::sql_evaluator::EvaluationNode;
 use super::VisitorContext;
+use crate::plan::Schema;
 use cubenativeutils::CubeError;
 use itertools::Itertools;
 use std::collections::HashMap;
 use std::rc::Rc;
 
 pub trait BaseMember {
-    fn to_sql(&self, context: Rc<VisitorContext>) -> Result<String, CubeError>;
+    fn to_sql(&self, context: Rc<VisitorContext>, schema: Rc<Schema>) -> Result<String, CubeError>;
     fn alias_name(&self) -> String;
     fn member_evaluator(&self) -> Rc<EvaluationNode>;
     fn full_name(&self) -> String {
         self.member_evaluator().full_name()
     }
     fn as_base_member(self: Rc<Self>) -> Rc<dyn BaseMember>;
+    fn cube_name(&self) -> &String;
+    fn name(&self) -> &String;
+    fn alias_suffix(&self) -> Option<String> {
+        None
+    }
 }
 
 pub struct BaseMemberHelper {}