+++ rust/cubesqlplanner/cubesqlplanner/src/planner/planners/mod.rs
@@ -3,6 +3,7 @@ pub mod full_key_query_aggregate_planner;
 pub mod join_planner;
 pub mod multi_stage;
 pub mod multi_stage_query_planner;
+pub mod multiplied_measures_query_planner;
 pub mod order_planner;
 pub mod simple_query_planer;
 
@@ -10,5 +11,6 @@ pub use common_utils::CommonUtils;
 pub use full_key_query_aggregate_planner::FullKeyAggregateQueryPlanner;
 pub use join_planner::JoinPlanner;
 pub use multi_stage_query_planner::MultiStageQueryPlanner;
+pub use multiplied_measures_query_planner::MultipliedMeasuresQueryPlanner;
 pub use order_planner::OrderPlanner;
 pub use simple_query_planer::SimpleQueryPlanner;